Split trakt service implementation to trakt-service module.
