+++ shared/data/trakt-service/api/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    id("tvmaniac.kmm.api")
+    alias(libs.plugins.serialization)
+}
+
+kotlin {
+    android()
+    ios()
+
+    sourceSets {
+        sourceSets["commonMain"].dependencies {
+            implementation(libs.ktor.serialization)
+        }
+    }
+}
+
+android {
+    namespace = "com.thomaskioko.tvmaniac.trakt.service.api"
+}
