+++ shared/data/trakt-service/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/service/api/model/ErrorResponse.kt
@@ -1,0 +1,9 @@
+package com.thomaskioko.tvmaniac.trakt.service.api.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ErrorResponse(
+    @SerialName("status_message") val message: String,
+)
