+++ shared/data/trakt-service/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/service/api/model/RefreshAccessTokenBody.kt
@@ -1,0 +1,13 @@
+package com.thomaskioko.tvmaniac.trakt.service.api.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RefreshAccessTokenBody(
+    @SerialName("refresh_token") val refreshToken: String?,
+    @SerialName("client_id") val clientId: String = "",
+    @SerialName("client_secret") val clientSecret: String = "",
+    @SerialName("redirect_uri") val redirectUri: String = "",
+    @SerialName("grant_type") val grantType: String = "refresh_token"
+)
