+++ shared/data/trakt-service/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/service/api/model/TraktSeasonsResponse.kt
@@ -1,0 +1,23 @@
+package com.thomaskioko.tvmaniac.trakt.service.api.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TraktSeasonsResponse(
+    @SerialName("number") val number: Int,
+    @SerialName("ids") val ids: SeasonIds,
+    @SerialName("votes") val votes: Int,
+    @SerialName("rating") val rating: Double,
+    @SerialName("aired_episodes") val airedEpisodes: Int,
+    @SerialName("episode_count") val episodeCount: Int,
+    @SerialName("title") val title: String,
+    @SerialName("overview") val overview: String?,
+    @SerialName("first_aired") val firstAirDate: String?,
+)
+
+@Serializable
+data class SeasonIds(
+    @SerialName("trakt") val trakt: Int,
+    @SerialName("tmdb") val tmdb: Int?,
+)
