+++ shared/data/trakt-service/implementation/src/commonMain/kotlin/com/thomaskioko/trakt/service/implementation/TraktServiceModule.kt
@@ -1,0 +1,5 @@
+package com.thomaskioko.trakt.service.implementation
+
+import org.koin.core.module.Module
+
+expect fun traktServiceModule(): Module
