--- shared/data/trakt/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/api/model/TraktAccessTokenResponse.kt
@@ -1,14 +1,0 @@
-package com.thomaskioko.tvmaniac.trakt.api.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TraktAccessTokenResponse(
-    @SerialName("scope") val scope: String?,
-    @SerialName("access_token") val accessToken: String?,
-    @SerialName("created_at") val createdAt: Long?,
-    @SerialName("expires_in") val expiresIn: Long?,
-    @SerialName("refresh_token") val refreshToken: String?,
-    @SerialName("token_type") val tokenType: String?
-)
