+++ shared/data/trakt/implementation/build.gradle.kts
@@ -1,9 +1,5 @@
-import com.codingfeline.buildkonfig.compiler.FieldSpec.Type.STRING
-
 plugins {
     id("tvmaniac.kmm.impl")
-    id("com.codingfeline.buildkonfig")
-    alias(libs.plugins.serialization)
 }
 
 kotlin {
@@ -13,9 +9,6 @@ kotlin {
     sourceSets {
         sourceSets["androidMain"].dependencies {
             implementation(project(":shared:core:util"))
-            implementation(project(":shared:data:network"))
-            implementation(libs.appauth)
-            implementation(libs.ktor.okhttp)
             implementation(libs.hilt.android)
             configurations["kapt"].dependencies.add(
                 org.gradle.api.internal.artifacts.dependencies.DefaultExternalModuleDependency(
@@ -28,22 +21,13 @@ kotlin {
 
 
         sourceSets["commonMain"].dependencies {
-            implementation(project(":shared:data:network"))
-            implementation(project(":shared:data:trakt:api"))
             implementation(project(":shared:data:category:api"))
+            implementation(project(":shared:data:trakt:api"))
+            implementation(project(":shared:data:trakt-service:api"))
             implementation(libs.sqldelight.extensions)
-            implementation(libs.ktor.core)
-            implementation(libs.ktor.negotiation)
-            implementation(libs.ktor.logging)
-            implementation(libs.ktor.serialization.json)
-            implementation(libs.ktor.serialization)
         }
 
         sourceSets["iosMain"].dependencies {
-            implementation(project(":shared:data:network"))
-            implementation(project(":shared:data:trakt:api"))
-            implementation(libs.ktor.logging)
-            implementation(libs.ktor.darwin)
         }
 
     }
@@ -51,32 +35,4 @@ kotlin {
 
 android {
     namespace = "com.thomaskioko.tvmaniac.trakt.auth.implementation"
-}
-
-buildkonfig {
-    packageName = "com.thomaskioko.tvmaniac.trakt.auth.implementation"
-
-    defaultConfigs {
-        buildConfigField(
-            STRING,
-            "TRAKT_CLIENT_ID",
-            "\"" + propOrDef("TRAKT_CLIENT_ID", "") + "\""
-        )
-        buildConfigField(
-            STRING,
-            "TRAKT_CLIENT_SECRET",
-            "\"" + propOrDef("TRAKT_CLIENT_SECRET", "") + "\""
-        )
-        buildConfigField(
-            STRING,
-            "TRAKT_REDIRECT_URI",
-            "\"" + propOrDef("TRAKT_REDIRECT_URI", "") + "\""
-        )
-    }
-}
-
-fun <T : Any> propOrDef(propertyName: String, defaultValue: T): T {
-    @Suppress("UNCHECKED_CAST")
-    val propertyValue = project.properties[propertyName] as T?
-    return propertyValue ?: defaultValue
 }
