--- shared/data/trakt/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/implementation/TraktHttpClient.kt
@@ -1,53 +1,0 @@
-package com.thomaskioko.tvmaniac.trakt.implementation
-
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.plugins.HttpTimeout
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.EMPTY
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.http.URLProtocol
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.json.Json
-
-
-fun traktHttpClient(
-    isDebug: Boolean = false,
-    json: Json,
-    httpClientEngine: HttpClientEngine,
-) = HttpClient(httpClientEngine) {
-    install(ContentNegotiation) {
-        json(json = json)
-    }
-
-    defaultRequest {
-        url {
-            protocol = URLProtocol.HTTPS
-            host = "api.trakt.tv"
-        }
-    }
-
-    install(HttpTimeout) {
-        requestTimeoutMillis = 60000
-        connectTimeoutMillis = 60000
-        socketTimeoutMillis = 60000
-    }
-
-
-    install(Logging) {
-        level = LogLevel.INFO
-        logger = if (isDebug) {
-            object : Logger {
-                override fun log(message: String) {
-                    co.touchlab.kermit.Logger.d { message }
-                }
-            }
-
-        } else {
-            Logger.EMPTY
-        }
-    }
-}
