+++ shared/data/trakt/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/implementation/TraktShowRepositoryImpl.kt
@@ -18,17 +18,17 @@ import com.thomaskioko.tvmaniac.category.api.model.Category.ANTICIPATED
 import com.thomaskioko.tvmaniac.category.api.model.Category.FEATURED
 import com.thomaskioko.tvmaniac.category.api.model.Category.POPULAR
 import com.thomaskioko.tvmaniac.category.api.model.Category.TRENDING
-import com.thomaskioko.tvmaniac.trakt.api.TraktService
 import com.thomaskioko.tvmaniac.trakt.api.TraktShowRepository
 import com.thomaskioko.tvmaniac.trakt.api.cache.TraktFollowedCache
 import com.thomaskioko.tvmaniac.trakt.api.cache.TraktUserCache
 import com.thomaskioko.tvmaniac.trakt.api.cache.TvShowCache
-import com.thomaskioko.tvmaniac.trakt.api.model.ErrorResponse
-import com.thomaskioko.tvmaniac.trakt.api.model.TraktShowResponse
 import com.thomaskioko.tvmaniac.trakt.implementation.mapper.responseToCache
 import com.thomaskioko.tvmaniac.trakt.implementation.mapper.showResponseToCacheList
 import com.thomaskioko.tvmaniac.trakt.implementation.mapper.showsResponseToCacheList
 import com.thomaskioko.tvmaniac.trakt.implementation.mapper.toCategoryCache
+import com.thomaskioko.tvmaniac.trakt.service.api.TraktService
+import com.thomaskioko.tvmaniac.trakt.service.api.model.ErrorResponse
+import com.thomaskioko.tvmaniac.trakt.service.api.model.TraktShowResponse
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
