--- shared/data/trakt/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/implementation/model/RefreshAccessTokenBody.kt
@@ -1,13 +1,0 @@
-package com.thomaskioko.tvmaniac.trakt.implementation.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RefreshAccessTokenBody(
-    @SerialName("refresh_token") val refreshToken: String?,
-    @SerialName("client_id") val clientId: String = "",
-    @SerialName("client_secret") val clientSecret: String = "",
-    @SerialName("redirect_uri") val redirectUri: String = "",
-    @SerialName("grant_type") val grantType: String = "refresh_token"
-)
